+++ examples/events.rs
@@ -1,8 +1,5 @@
 use macroquad::megaui::{widgets, Vector2};
-use macroquad::{
-    draw_window, hash, is_key_down, is_mouse_button_down, mouse_position, mouse_wheel, next_frame,
-    MouseButton, Vec2,
-};
+use macroquad::prelude::*;
 
 #[macroquad::main("Events")]
 async fn main() {
