+++ megaui/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "megaui"
-version = "0.2.12"
+version = "0.2.13"
 authors = ["fl3 <not.fl3@gmail.com>"]
 edition = "2018"
 license = "MIT/Apache-2.0"
