+++ quad-gl/Cargo.toml
@@ -6,9 +6,9 @@ homepage = "https://github.com/not-fl3/quad-gl"
 license = "MIT/Apache-2.0"
 name = "quad-gl"
 repository = "https://github.com/not-fl3/quad-gl"
-version = "0.2.4"
+version = "0.2.5"
 
 [dependencies]
-miniquad = { git = "https://github.com/not-fl3/miniquad.git", branch = "unstable" }
+miniquad = "0.3.0-alpha.15"
 glam = {version = "0.9", features = ["scalar-math"] }
 image = { version = "0.22", default-features = false, features = ["png_codec"] }
