+++ src/material.rs
@@ -1,18 +1,27 @@
-use super::*;
+//! Custom materials - shaders, uniforms. 
 
-///
+use quad_gl::GlPipeline;
+use crate::get_context;
+use miniquad::{PipelineParams, UniformType, ShaderError};
+
+/// Material instance loaded on GPU. 
 #[derive(Copy, Clone, Debug, PartialEq)]
 pub struct Material {
     pipeline: GlPipeline,
 }
 
 impl Material {
+    /// Set GPU uniform valuye for this material.
+    /// "name" should be from "uniforms" list used for material creation.
+    /// Otherwise uniform value would be silently ignored.
     pub fn set_uniform<T>(&self, name: &str, uniform: T) {
         let context = &mut get_context().draw_context;
 
         context.gl.set_uniform(self.pipeline, name, uniform);
     }
 
+    /// Delete this material. Using deleted material for either rendering
+    /// or uniforms manipulation will result internal GL errors.
     pub fn delete(&mut self) {
         let context = &mut get_context().draw_context;
 
@@ -20,6 +29,9 @@ impl Material {
     }
 }
 
+/// Params used for material loading.
+/// It is not possible to change material params at runtime, so this
+/// struct is used only once - at "load_material".
 pub struct MaterialParams {
     /// miniquad pipeline configuration for this material.
     /// Things like blending, culling, depth dest
@@ -56,12 +68,14 @@ pub fn load_material(
     Ok(Material { pipeline })
 }
 
+/// Al followed macroquad rendering calls will use the given material.
 pub fn gl_use_material(material: Material) {
     let context = &mut get_context().draw_context;
 
     context.gl.pipeline(Some(material.pipeline));
 }
 
+/// Use default macroquad material. 
 pub fn gl_use_default_material() {
     let context = &mut get_context().draw_context;
